fix: 로그인 시 세션 저장하는 로직 추가 및 회원가입 시 세션 저장하는 로직 제거
