+++ tomcat/src/main/java/nextstep/jwp/handler/LoginHandler.java
@@ -1,10 +1,8 @@
 package nextstep.jwp.handler;
 
-import java.util.Objects;
 import java.util.Optional;
 import nextstep.Application;
 import nextstep.jwp.db.InMemoryUserRepository;
-import nextstep.jwp.exception.UserNotFoundException;
 import nextstep.jwp.model.User;
 import org.apache.catalina.Manager;
 import org.apache.catalina.Session;
@@ -14,6 +12,7 @@ import org.apache.coyote.http11.HttpRequestHeader;
 import org.apache.coyote.http11.HttpResponse;
 import org.apache.coyote.http11.enums.HttpMethod;
 import org.apache.coyote.http11.enums.HttpStatus;
+import org.apache.coyote.http11.utils.UuidUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -34,34 +33,42 @@ public class LoginHandler {
             return HttpResponse.of(httpRequest, HttpStatus.OK, "/login.html");
         }
 
-        final HttpRequestHeader requestHeader = httpRequest.getHttpRequestHeader();
         final HttpRequestBody requestBody = httpRequest.getHttpRequestBody();
-        final User user = findUser(requestHeader, requestBody);
+        final Optional<User> findUser = findUser(requestBody);
 
+        if (findUser.isEmpty()) {
+            return HttpResponse.of(httpRequest, HttpStatus.UNAUTHORIZED, "/401.html");
+        }
+
+        final User user = findUser.get();
         final String password = requestBody.findByKey(PASSWORD);
         if (user.checkPassword(password)) {
             log.info(user.toString());
+            setUpSession(user, httpRequest.getHttpRequestHeader());
             return generateSuccessResponse(httpRequest);
         }
 
         return HttpResponse.of(httpRequest, HttpStatus.UNAUTHORIZED, "/401.html");
     }
 
-    private User findUser(final HttpRequestHeader requestHeader, final HttpRequestBody requestBody) {
-        Optional<String> jSessionId = requestHeader.findJSessionId();
+    private void setUpSession(final User user, final HttpRequestHeader httpRequestHeader) {
+        Optional<String> jSessionId = httpRequestHeader.findJSessionId();
         if (jSessionId.isEmpty()) {
-            final String account = requestBody.findByKey(ACCOUNT);
-            return InMemoryUserRepository.findByAccount(account)
-                    .orElseThrow(() -> new UserNotFoundException(account));
+            addSession(user, UuidUtil.randomUuidString());
+            return;
         }
+        addSession(user, jSessionId.get());
+    }
 
-        final String id = jSessionId.get();
-        final Session session = manager.findSession(id);
-        if (Objects.isNull(session)) {
-            throw new UserNotFoundException(id);
-        }
+    private void addSession(final User user, final String jSessionId) {
+        Session session = new Session(jSessionId);
+        session.addUser(user);
+        manager.add(session);
+    }
 
-        return session.getUser();
+    private Optional<User> findUser(final HttpRequestBody requestBody) {
+        final String account = requestBody.findByKey(ACCOUNT);
+        return InMemoryUserRepository.findByAccount(account);
     }
 
     private HttpResponse generateSuccessResponse(final HttpRequest httpRequest) {
